+++ src/webdata/cesiumTileset/cesiumTilesetApi.ts
@@ -1,6 +1,6 @@
 import { CesiumTilesetLayer } from '@vcmap/core';
 import { type VcsUiApp } from '@vcmap/ui';
-import { reactive } from 'vue';
+import { reactive, toRaw } from 'vue';
 import { getTreeviewDefaultActions } from '../webdataActionsHelper.js';
 import type { DataItem } from '../webdataConstants.js';
 import { WebdataTypes } from '../webdataConstants.js';
@@ -46,6 +46,7 @@ export function itemToCesiumTileset(item: DataItem): CesiumTilesetLayer {
     url: item.url,
     name: item.name,
     activeOnStartup: true,
+    headers: toRaw(item.headers),
     properties: { title: item.title },
   });
 }