+++ src/webdata/pointcloud/pointCloudApi.ts
@@ -1,6 +1,6 @@
 import { PointCloudLayer } from '@vcmap/core';
 import { type VcsUiApp } from '@vcmap/ui';
-import { reactive } from 'vue';
+import { reactive, toRaw } from 'vue';
 import { getTreeviewDefaultActions } from '../webdataActionsHelper.js';
 import type { DataItem } from '../webdataConstants.js';
 import { WebdataTypes } from '../webdataConstants.js';
@@ -43,7 +43,7 @@ export function itemToPointCloud(item: DataItem): PointCloudLayer {
   return new PointCloudLayer({
     url: item.url,
     name: item.name,
-    activeOnStartup: true,
+    headers: toRaw(item.headers),
     properties: { title: item.title },
   });
 }